fix(ui): fix color parsing and add block management
- Cast room color to Int to resolve Compose compilation error
- Add support for creating and deleting time blocks in RoomScreen
- Implement addBlock and deleteBlock logic in ViewModel

+++ app/src/main/java/com/example/ui/screens/HomeScreen.kt
@@ -228,7 +228,7 @@ fun HomeScreen(
 
 @Composable
 fun RoomCard(room: RoomEntity, onClick: () -> Unit) {
-    val roomColor = Color(room.colorArgb)
+    val roomColor = Color(room.colorArgb.toInt())
     val interactionSource = remember { MutableInteractionSource() }
     val isPressed by interactionSource.collectIsPressedAsState()
